move displayConfig to utils, needed by processor-plugin
