+++ cmd/conduit/root/connectorplugins/describe.go
@@ -17,10 +17,7 @@ package connectorplugins
 import (
 	"context"
 	"fmt"
-	"strings"
 
-	"github.com/alexeyco/simpletable"
-	configv1 "github.com/conduitio/conduit-commons/proto/config/v1"
 	"github.com/conduitio/conduit/cmd/conduit/api"
 	"github.com/conduitio/conduit/cmd/conduit/cecdysis"
 	"github.com/conduitio/conduit/cmd/conduit/internal"
@@ -96,99 +93,10 @@ func displayConnectorPluginsDescription(c *apiv1.ConnectorPluginSpecifications) 
 	fmt.Printf("Version: %s\n", c.Version)
 	if len(c.SourceParams) > 0 {
 		fmt.Printf("\nSource Parameters:\n")
-		displayConnectorPluginParams(c.SourceParams)
+		internal.DisplayConfigParams(c.SourceParams)
 	}
 	if len(c.DestinationParams) > 0 {
 		fmt.Printf("\nDestination Parameters:\n")
-		displayConnectorPluginParams(c.DestinationParams)
+		internal.DisplayConfigParams(c.DestinationParams)
 	}
-}
-
-func displayConnectorPluginParams(cfg map[string]*configv1.Parameter) {
-	table := simpletable.New()
-
-	table.Header = &simpletable.Header{
-		Cells: []*simpletable.Cell{
-			{Align: simpletable.AlignCenter, Text: "NAME"},
-			{Align: simpletable.AlignCenter, Text: "TYPE"},
-			{Align: simpletable.AlignCenter, Text: "DESCRIPTION"},
-			{Align: simpletable.AlignCenter, Text: "DEFAULT"},
-			{Align: simpletable.AlignCenter, Text: "VALIDATIONS"},
-		},
-	}
-
-	// create slices for ordered parameters, needed to keep the name
-	var requiredParams, otherParams, sdkParams []struct {
-		name  string
-		param *configv1.Parameter
-	}
-
-	// separate parameters into three groups for ordering purposes
-	for name, param := range cfg {
-		switch {
-		case strings.HasPrefix(name, "sdk"):
-			sdkParams = append(sdkParams, struct {
-				name  string
-				param *configv1.Parameter
-			}{name: name, param: param})
-		case isRequired(param.Validations):
-			requiredParams = append(requiredParams, struct {
-				name  string
-				param *configv1.Parameter
-			}{name: name, param: param})
-		default:
-			otherParams = append(otherParams, struct {
-				name  string
-				param *configv1.Parameter
-			}{name: name, param: param})
-		}
-	}
-
-	// combine ordered parameters
-	orderedParams := append(requiredParams, otherParams...) //nolint:gocritic // intentional
-	orderedParams = append(orderedParams, sdkParams...)
-
-	for _, item := range orderedParams {
-		r := []*simpletable.Cell{
-			{Align: simpletable.AlignLeft, Text: item.name},
-			{Align: simpletable.AlignLeft, Text: formatType(item.param.GetType().String())},
-			{Align: simpletable.AlignLeft, Text: internal.FormatLongString(item.param.Description, 100)},
-			{Align: simpletable.AlignLeft, Text: item.param.Default},
-			{Align: simpletable.AlignLeft, Text: formatValidations(item.param.Validations)},
-		}
-		table.Body.Cells = append(table.Body.Cells, r)
-	}
-
-	table.SetStyle(simpletable.StyleDefault)
-	fmt.Println(table.String())
-}
-
-func formatType(input string) string {
-	return strings.TrimPrefix(strings.ToLower(input), "type_")
-}
-
-func isRequired(validations []*configv1.Validation) bool {
-	for _, validation := range validations {
-		if strings.ToUpper(validation.GetType().String()) == configv1.Validation_TYPE_REQUIRED.String() {
-			return true
-		}
-	}
-	return false
-}
-
-func formatValidations(v []*configv1.Validation) string {
-	var result strings.Builder
-	for _, validation := range v {
-		if result.Len() > 0 {
-			result.WriteString(", ")
-		}
-		formattedType := formatType(validation.GetType().String())
-		value := validation.GetValue()
-		if value == "" {
-			result.WriteString(fmt.Sprintf("[%s]", formattedType))
-		} else {
-			result.WriteString(fmt.Sprintf("[%s=%s]", formattedType, value))
-		}
-	}
-	return result.String()
 }
